Reorganize actions by resource
