+++ app/webpack/actions/matches.js
@@ -1,0 +1,3 @@
+import { createAction } from 'redux-actions';
+
+export const updateMatch = createAction('MATCH_UPDATE')
